refactor: add kotlin-test dependency and change the packaging of ApiInterceptor
