+++ remote_datasource/src/main/java/com/baghdad/remote_datasource/interceptor/ApiInterceptor.kt
@@ -1,25 +1,26 @@
-package com.baghdad.remote_datasource
+package com.baghdad.remote_datasource.interceptor
 
-import io.ktor.client.*
-import io.ktor.client.plugins.*
-import io.ktor.client.request.*
-import io.ktor.util.*
+import io.ktor.client.HttpClient
+import io.ktor.client.plugins.HttpClientPlugin
+import io.ktor.client.request.HttpRequestPipeline
+import io.ktor.client.request.parameter
+import io.ktor.util.AttributeKey
 
 class ApiInterceptor(private val apiKey: String) {
     class Config {
         var apiKey: String = ""
     }
-    
+
     companion object Plugin : HttpClientPlugin<Config, ApiInterceptor> {
         override val key: AttributeKey<ApiInterceptor> = AttributeKey("ApiKeyInterceptor")
-        
+
         override fun prepare(block: Config.() -> Unit): ApiInterceptor {
             val config = Config().apply(block)
             return ApiInterceptor(config.apiKey)
         }
-        
+
         override fun install(plugin: ApiInterceptor, scope: HttpClient) {
-            scope.requestPipeline.intercept(HttpRequestPipeline.State) {
+            scope.requestPipeline.intercept(HttpRequestPipeline.Phases.State) {
                 context.parameter("api_key", plugin.apiKey)
             }
         }
